lab4_v3/requirements.txt
attrs==25.4.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
flasgger==0.9.7.1
Flask==2.3.3
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mistune==3.1.4
packaging==25.0
pycparser==2.23
PyMySQL==1.1.0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
typing_extensions==4.15.0
Werkzeug==3.1.4
